refactor to separate pages
